pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mapfree"
version = "1.1.0"
description = "Modular photogrammetry pipeline engine (CLI + Qt GUI)"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.10"
authors = [{ name = "MapFree Contributors" }]
keywords = ["photogrammetry", "colmap", "3d-reconstruction", "pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
# PySide6 must be installed via conda-forge (see environment.yml), not pip, to avoid Qt DLL conflicts with COLMAP/GDAL.
dependencies = [
    "PyYAML>=6.0",
    "psutil>=5.9.0",
    "numpy>=1.20.0",
    "matplotlib>=3.5.0",
    "Pillow>=9.0.0",
    "exifread>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "flake8>=6.0",
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "pytest-cov>=4.0",
    "pyinstaller>=6.0",
]
viewer = [
    "pyqtgraph>=0.13.0",
]

[project.scripts]
mapfree = "mapfree.__main__:main"
pytest = "pytest:main"

[tool.setuptools.packages.find]
include = ["mapfree*"]

[tool.coverage.run]
omit = [
    "mapfree/gui/*",
    "mapfree/viewer/*",
]

[tool.pytest.ini_options]
markers = [
    "gui: GUI tests (require DISPLAY on Linux; run with: xvfb-run pytest tests/gui)",
]
testpaths = ["tests"]
